Try fixing health bugs

+++ src/constants.rs
@@ -8,7 +8,7 @@ pub const WORLD_SIZE: f32 = 3000.;
 pub const MAX_SPEED: f32 = 400.;
 pub const MIN_SPEED: f32 = 50.;
 pub const DEFAULT_ACCELERATION: f32 = 200.;
-pub const DEFAULT_HEALTH: u8 = 150;
+pub const DEFAULT_HEALTH: i16 = 150;
 
 pub const DEFAULT_AGILITY: f32 = 1.;
 pub const ANGULAR_FADE: f32 = 0.9;
@@ -28,6 +28,6 @@ pub const SCANCODE_S: u32 = 31;
 pub const SCANCODE_D: u32 = 32;
 
 pub const POWERUP_AMOUNT: u32 = 5;
-pub const POWERUP_HEALTH_BOOST: u8 = 40;
+pub const POWERUP_HEALTH_BOOST: i16 = 40;
 
 pub const MINI_MAP_SIZE: f32 = 300.;
